requirements.txt
# GeoNode
-e git+https://github.com/GeoNode/geonode.git@master#egg=geonode

html5lib==0.9999999
djangorestframework>=3.1.3,<3.7
wagtail==1.13.4
# django-filter should be already in version 2.0.0; need a previous version for wagtail 1.13.4
django-filter==1.1.0
puput==0.9.2
django_compressor
django-social-share
django-widget-tweaks
